Consolidate the bucardo index & constraint logic better.
The previous commit created some duplicate code in order to handle dropping and
recreating of indexes and constraints. This commit consolidates the code
better. In the process, it creates an extra database column to store the DDL
for dropping the indexes and constraints. The previous iteration had only one
DDL column, and it was for recreating the objects.