INF-6236: Incorporate async kicking changes into the retry plugin.
A recent commit added new functionality to the bucardo plugin, relating to
sync-kicking behavior. It also changed the name of a function. Since the retry
plugin inherits from the bucardo plugin and reuses some of its logic, the
corresponding changes need to be applied to the retry plugin.

There was also a bit of verbiage that dates to a very old version of this code
that needed updating.