Fix trigger logic bugs.
Bug 1: In a cascading replication toplogy, truncates on the first database
were not being propagated down to the third database. Setting the trunc trigger
on the second database to always be enabled fixes this.

Bug 2: Enabling and disabling of triggers wasn't idempotent and was
furthermore somewhat arbitrary if you had a certain combination of settings.
This change makes the trigger logic consistent with the intent of the settings.

Bug 3: A function was called by an old name that got missed in a previous refactor.

Bug 4: The documentation referred to the old state in a couple places.

Also updated some double quotes to single quotes for consistency.